Boilerplate for the divert converter (#66)
There aer two steps to building a training file. The first is to convert the `DiVertAnalysisR21` files into something that can be used by the tools. The second is to combine those files into a full training sample. This PR addresses the first of those two steps.

* Added the command to `convert divertanalysis` files.
* Supports wildcards
* Added code that will convert the file as `sig`, `qcd`, or `bib`.
* Output will be written to whatever directory is specified.
* BIB conversion does not yet work (see #67).
* Updated README file

Convert DiVertAnalysis files to input training files 
Fixes #65