Only plot signal files that are signal file events (#182)
We were making an assumption that any event with a non-zero $m_S$ or $m_H$ was signal. Turns out not to be the case - Felix's old trianing files seem to have assigned $m_H$ and $m_S$ values to the jet data!

This first filters on the `label` before looking at `llp_mS` or `llp_mH`.

Fixes #172